移动cache到common模块
