+++ manager/src/main/java/org/dromara/hertzbeat/manager/service/impl/NoticeConfigServiceImpl.java
@@ -17,10 +17,10 @@
 
 package org.dromara.hertzbeat.manager.service.impl;
 
+import org.dromara.hertzbeat.common.cache.CacheFactory;
+import org.dromara.hertzbeat.common.cache.ICacheService;
 import org.dromara.hertzbeat.common.entity.alerter.Alert;
 import org.dromara.hertzbeat.common.util.CommonConstants;
-import org.dromara.hertzbeat.manager.cache.CacheFactory;
-import org.dromara.hertzbeat.manager.cache.ICacheService;
 import org.dromara.hertzbeat.manager.component.alerter.DispatcherAlarm;
 import org.dromara.hertzbeat.manager.dao.NoticeReceiverDao;
 import org.dromara.hertzbeat.manager.dao.NoticeRuleDao;
